resolving -Wall compiler warnings

+++ src/crorc_qsfp_ctrl.cpp
@@ -39,7 +39,7 @@ float mWatt2dBm(float mwatt) { return 10 * log10(mwatt); }
 
 string stripWhitespaces(string input) {
   string output;
-  for (int i = 0; i < input.length(); i++) {
+  for (size_t i = 0; i < input.length(); i++) {
     int c = input[i];
     if (c != 32) {
       output += (input[i]);
